Move layer trait related types to the layer_trait module
